Own Genetic mutation selection and replacement admission
Use canonical marked CellClass membership for per-cell launch, excluding factory-held and unmarked objects. Native selection rechecked at 006CD954-006CD9D6 in retail gamemd SHA256 1CDD1180E49024FBDA8AD568CAAC2E86E856063FF67AB38F62B7D2C7BB84298C. Preserve legacy stable-ID batch, damage and immediate BRUTE policy behind one private completed operation. Terminal Infantry disposition remains separate and open.
